+++ src/codemagic/tools/_app_store_connect/abstract_base_action.py
@@ -54,6 +54,17 @@ class AbstractBaseAction(ResourceManagerMixin, PathFinderMixin, metaclass=ABCMet
             should_print: bool = True) -> BetaBuildLocalization:
         ...
 
+    def add_beta_test_info(self,
+                           build_id: ResourceId,
+                           beta_build_localizations: Types.BetaBuildLocalizations,
+                           locale: Optional[Locale],
+                           whats_new: Optional[Union[str, Types.WhatsNewArgument]] = None):
+        ...
+
+    def submit_to_testflight(
+            self, build_id: ResourceId, max_build_processing_wait: Optional[Types.MaxBuildProcessingWait] = None):
+        ...
+
     def list_apps(self,
                   bundle_id_identifier: Optional[str] = None,
                   application_id: Optional[ResourceId] = None,